+++ src/fable-library-ts/Native.fs
@@ -96,3 +96,6 @@ module Helpers =
     let sortInPlaceWithImpl (comparer: 'T -> 'T -> int) (array: 'T[]) : unit = nativeOnly
 
     let inline arrayAccessImpl (index: int) (array: 'T[]) : 'T = emitJsExpr (index, array) "$1[$0]"
+
+    let inline arraySetItemImpl (array: 'T[]) (index: int) (item: 'T) : unit =
+        emitJsExpr (array, index, item) "$0[$1] = $2"