+++ src/Fable.Cli/CHANGELOG.md
@@ -19,6 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Update to Fable.AST 4.4.0
 
+### Fixed
+
+#### JavaScript
+
+* [GH-3748](https://github.com/fable-compiler/Fable/pull/3748) Accessing an array out of bounds should emit an exception (by @MangelMaxime)
+* [GH-3748](https://github.com/fable-compiler/Fable/pull/3748) Setting an array out of bounds should emit an exception (by @MangelMaxime)
+
 ## 4.12.1 - 2024-02-13
 
 ### Fixed